refactor method into util file
